+++ packages/colorpickers/src/styled/ColorPicker/StyledSliders.ts
@@ -5,7 +5,7 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-import styled, { css } from 'styled-components';
+import styled, { DefaultTheme, ThemeProps } from 'styled-components';
 import { retrieveComponentStyles, DEFAULT_THEME } from '@zendeskgarden/react-theming';
 import { getTrackHeight, getTrackMargin } from '../common/StyledRange';
 
@@ -14,6 +14,29 @@ const COMPONENT_ID = 'colorpickers.colorpicker_sliders';
 interface IStyledSlidersProps {
   $isOpaque?: boolean;
 }
+
+const sizeStyles = (props: IStyledSlidersProps & ThemeProps<DefaultTheme>) => {
+  if (props.$isOpaque) return undefined;
+
+  const trackHeight = getTrackHeight(props);
+  const trackMargin = getTrackMargin(props);
+
+  return `
+    & > * {
+      position: absolute;
+      width: 100%;
+      height: ${trackMargin * 2 + trackHeight}px;
+    }
+
+    & > :first-child {
+      top: -${trackMargin}px;
+    }
+
+    & > :last-child {
+      bottom: -${trackMargin}px;
+    }
+  `;
+};
 
 export const StyledSliders = styled.div.attrs({
   'data-garden-id': COMPONENT_ID,
@@ -24,24 +47,7 @@ export const StyledSliders = styled.div.attrs({
   margin-${props => (props.theme.rtl ? 'right' : 'left')}: ${props => props.theme.space.base * 2}px;
   width: 100%;
 
-  ${props =>
-    !props.$isOpaque &&
-    css`
-      & > * {
-        position: absolute;
-        width: 100%;
-        height: ${getTrackMargin(props) * 2 + getTrackHeight(props)}px;
-      }
-
-      & > :first-child {
-        top: -${getTrackMargin}px;
-      }
-
-      & > :last-child {
-        bottom: -${getTrackMargin}px;
-      }
-    `}
-  
+  ${sizeStyles}
 
   ${props => retrieveComponentStyles(COMPONENT_ID, props)};
 `;
